refactor: scope the app under (app) so /studio inherits a bare root layout
The root layout is now a bare shell — html/body, global CSS, font
variables — and everything app-flavored (providers, metadata, JSON-LD,
toasts, analytics, the satus-version script) moves into
app/(app)/layout.tsx, a route group every route except /studio lives
under. Anything added to the root layout from now on is a deliberate
decision to ship it to Studio too.

- Moved all routes under app/(app)/ (pure renames); not-found recreated
  inside the group with a [...unmatched] catch-all so 404s render with
  the app providers (static segments /studio and /api resolve first).
- Font variables stay on <html> in the root layout — portaled UI
  (toasts, dialogs) inherits them for free; Studio paying a font preload
  is an acceptable cost.
- ReactTempus keeps patch={!isDraftMode} — the prop defaults to true in
  tempus 1.0.0, and patching rAF in draft mode conflicts with Visual
  Editing.
- /studio page now awaits connection() behind its own Suspense boundary:
  outside the (app) group its tree reads no dynamic API, so Cache
  Components would otherwise fully prerender NextStudio at build time,
  which crashes. This restores main's request-time behavior.
- global-error.tsx imports the stylesheet itself (it replaces the root
  layout, taking the CSS import with it) and drops Wrapper so the
  last-resort crash boundary no longer depends on Lenis/WebGL/theme
  chrome that may be the thing that just crashed.
- Updated setup-project / integration-bundles paths, tests, and docs.
