+++ app/(app)/loading.tsx
@@ -1,0 +1,10 @@
+// No <Wrapper> here: with cacheComponents, this group-level loading fallback
+// must be statically renderable — Wrapper mounts <Theme>, which reads
+// uncached data and fails prerendering. Keep this fallback dependency-free.
+export default function Loading() {
+  return (
+    <div className="flex min-h-dvh flex-col items-center justify-center font-mono uppercase">
+      <p>Cooking...</p>
+    </div>
+  )
+}
