+++ app/(app)/not-found.tsx
