+++ app/README.md
@@ -6,16 +6,24 @@ Next.js App Router pages and routes.
 
 ```
 app/
-├── page.tsx              # Manual / in-app documentation landing page
-├── layout.tsx            # Root layout
+├── layout.tsx            # Bare html/body shell shared with /studio
+├── global-error.tsx      # Root-level error boundary (wraps html+body)
+├── (app)/                # Every route EXCEPT /studio lives in this group
+│   ├── layout.tsx        # App providers, metadata, analytics
+│   ├── page.tsx          # Manual / in-app documentation landing page
+│   ├── loading.tsx       # App loading fallback
+│   ├── error.tsx         # Error boundary (thin wrapper over components/ui/error-view)
+│   ├── not-found.tsx     # 404 page
+│   └── [...unmatched]/   # Catch-all routing unmatched URLs to the 404 above
 ├── api/
 │   ├── draft-mode/       # Sanity draft mode
 │   └── revalidate/       # Webhook endpoint
-├── loading.tsx           # Global loading
-├── error.tsx             # Error boundary (thin wrapper over components/ui/error-view)
-├── global-error.tsx      # Root-level error boundary (wraps html+body)
-└── not-found.tsx         # 404 page
+└── studio/               # Sanity Studio — inherits only the bare root shell
 ```
+
+The root layout stays a bare shell on purpose: anything added to it is a
+deliberate decision to ship it to `/studio` too. App-flavored concerns
+(providers, metadata, analytics) belong in `app/(app)/layout.tsx`.
 
 ## Getting Started
 
@@ -27,8 +35,8 @@ bun run setup:project       # Choose what to keep/remove
 
 Then:
 
-1. Customize `page.tsx` for your homepage
-2. Add routes as folders with `page.tsx`
+1. Customize `(app)/page.tsx` for your homepage
+2. Add routes as folders with `page.tsx` inside `(app)/`
 
 ## Page Pattern
