+++ app/layout.tsx
@@ -1,100 +1,20 @@
-import { Analytics } from '@vercel/analytics/next'
-import { TransformProvider } from 'hamo'
-import type { Metadata, Viewport } from 'next'
-import { VisualEditing } from 'next-sanity/visual-editing'
-import { draftMode } from 'next/headers'
-import Script from 'next/script'
-import { type PropsWithChildren, Suspense } from 'react'
-import { ReactTempus } from 'tempus/react'
+import type { PropsWithChildren } from 'react'
 
-import { Link } from '@/components/ui/link'
-import { RealViewport } from '@/components/ui/real-viewport'
-import { ToastProvider, ToastViewport } from '@/components/ui/toast'
-import { APP_BASE_URL, env } from '@/lib/env'
-import { OptionalFeatures } from '@/lib/features'
-import { isConfigured } from '@/lib/integrations/registry'
-import { SanityLive } from '@/lib/integrations/sanity/live'
-import { JsonLd } from '@/lib/seo/json-ld'
-import { organizationSchema, websiteSchema } from '@/lib/seo/schemas'
-import { themes } from '@/lib/styles/colors'
 import { fontsVariable } from '@/lib/styles/fonts'
-import AppData from '@/package.json'
 
 import '@/lib/styles/css/index.css'
 
-const APP_NAME = AppData.name
-const APP_DEFAULT_TITLE = 'Satūs'
-const APP_TITLE_TEMPLATE = '%s - Satūs'
-const APP_DESCRIPTION = AppData.description
+/*
+  Bare shell shared by the app and /studio. Everything app-specific —
+  providers, metadata, JSON-LD, analytics, the satus-version script — lives
+  in app/(app)/layout.tsx so /studio doesn't inherit it. Anything added here
+  is a deliberate decision to ship it to Studio too.
 
-export const metadata: Metadata = {
-  metadataBase: new URL(APP_BASE_URL),
-  applicationName: APP_NAME,
-  title: {
-    default: APP_DEFAULT_TITLE,
-    template: APP_TITLE_TEMPLATE,
-  },
-  description: APP_DESCRIPTION,
-  alternates: {
-    canonical: '/',
-    languages: {
-      'en-US': '/en-US',
-    },
-    // Advertises the plain-text mirror via <link rel="alternate" type="text/plain">.
-    types: {
-      'text/plain': [{ url: '/llms.txt', title: 'llms.txt' }],
-    },
-  },
-  appleWebApp: {
-    capable: true,
-    statusBarStyle: 'default',
-    title: APP_DEFAULT_TITLE,
-  },
-  formatDetection: { telephone: false },
-  openGraph: {
-    type: 'website',
-    siteName: APP_NAME,
-    title: {
-      default: APP_DEFAULT_TITLE,
-      template: APP_TITLE_TEMPLATE,
-    },
-    description: APP_DESCRIPTION,
-    url: APP_BASE_URL,
-    images: [
-      {
-        url: '/opengraph-image.jpg',
-        width: 1200,
-        height: 630,
-        alt: APP_DEFAULT_TITLE,
-      },
-    ],
-    locale: 'en_US',
-  },
-  twitter: {
-    card: 'summary_large_image',
-    title: {
-      default: APP_DEFAULT_TITLE,
-      template: APP_TITLE_TEMPLATE,
-    },
-    description: APP_DESCRIPTION,
-  },
-  authors: [
-    { name: 'darkroom.engineering', url: 'https://darkroom.engineering' },
-  ],
-  ...(env.NEXT_PUBLIC_FACEBOOK_APP_ID
-    ? { other: { 'fb:app_id': env.NEXT_PUBLIC_FACEBOOK_APP_ID } }
-    : {}),
-}
-
-export const viewport: Viewport = {
-  themeColor: themes.red.primary,
-  colorScheme: 'normal',
-}
-
-export default async function Layout({ children }: PropsWithChildren) {
-  const { isEnabled: isDraftMode } = await draftMode()
-  const sanityConfigured = isConfigured('sanity')
-
+  Font variables stay here on <html> so portaled UI (toasts, dialogs…)
+  inherits them for free; /studio paying a font preload is an acceptable
+  cost for that simplicity.
+*/
+export default function Layout({ children }: PropsWithChildren) {
   return (
     <html
       lang="en"
@@ -107,60 +27,7 @@ export default async function Layout({ children }: PropsWithChildren) {
       // otherwise trip a hydration warning.
       suppressHydrationWarning
     >
-      {/* this helps to track Satus usage thanks to Wappalyzer */}
-      <Script
-        id="satus-version"
-        async
-      >{`window.satusVersion = '${AppData.version}';`}</Script>
-      <body>
-        {/* Entity identity for search and answer engines, on every page — deep
-            pages are landed on directly far more often than the homepage. */}
-        <JsonLd data={organizationSchema()} />
-        <JsonLd data={websiteSchema()} />
-        {/* Skip link for keyboard navigation accessibility */}
-        <Suspense fallback={null}>
-          <Link
-            href="#main-content"
-            className="focus:top-4 focus:left-4 focus:rounded focus:px-4 focus:py-2 sr-only focus:not-sr-only focus:fixed focus:z-9999 focus:bg-black focus:text-white focus:ring-2 focus:ring-white focus:outline-none"
-          >
-            Skip to main content
-          </Link>
-        </Suspense>
-        {/* Critical: CSS custom properties needed for layout */}
-        <RealViewport>
-          <ToastProvider>
-            <TransformProvider>
-              {/*
-                DO NOT add Header or Footer here.
-                They are included in the <Wrapper> component used by each page.
-                See: components/layout/wrapper/index.tsx
-              */}
-              {children}
-            </TransformProvider>
-            <ToastViewport />
-          </ToastProvider>
-        </RealViewport>
-        {/* Optional features - conditionally loaded based on configuration */}
-        <OptionalFeatures />
-
-        {/* Sanity Live - renders unconditionally when Sanity is configured for real-time updates.
-            includeDrafts subscribes the event stream to draft mutations so
-            Presentation-tool edits push to the preview without a manual refresh. */}
-        {sanityConfigured && <SanityLive includeDrafts={isDraftMode} />}
-
-        {/* Sanity Visual Editing - only when draft mode is enabled */}
-        {sanityConfigured && isDraftMode && (
-          <Suspense fallback={null}>
-            <VisualEditing />
-          </Suspense>
-        )}
-
-        {/* RAF management - lightweight, but don't patch in draft mode to avoid conflicts */}
-        <ReactTempus patch={!isDraftMode} />
-        {/* Vercel-hosted deployments only — the injected /_vercel/insights
-            script 404s on self-hosted or CI `next start`. */}
-        {process.env.VERCEL_ENV && <Analytics />}
-      </body>
+      <body>{children}</body>
     </html>
   )
 }
